+++ lua/true-zen/services/bottom/integrations/integration_vim_airline.lua
